requirements.txt

Flask == 1.1.2
Flask_Admin == 1.5.7
Flask_AppBuilder == 3.2.0
Flask_SQLAlchemy == 2.4.4
Pillow == 8.0.1
SQLAlchemy == 1.3.20
face_recognition == 1.3.0
imutils == 0.5.4
numpy == 1.19.2
opencv_contrib_python == 4.5.1.48
opencv_python == 4.5.1.48